+++ app/components/timeline/timeline-directives.js
@@ -190,7 +190,7 @@ angular.module('lizard-nxt')
               timelineLayers.events.layers.push(layer);
               timelineLayers.events.slugs.push(layer.slug);
             } else if (layer.format === "Store" &&
-                       layer.slug === "radar/basic") {
+                       layer.slug === "rain") {
               timelineLayers.rain = layer;
             }
           });