+++ CHANGES.rst
@@ -6,6 +6,8 @@ Unreleased (1.2.7) (XXXX-XX-XX)
 
 - Get colormap per aggWindow for rain.
 
+- Fix radar/basic slug confusion, store slug is now `rain`.
+
 
 Release 1.2.10 (2015-1-22)
 ---------------------